Code Refactored

+++ RentACarSpring/src/main/java/com/etiya/rentACarSpring/businnes/concretes/CarManager.java
@@ -3,8 +3,12 @@ package com.etiya.rentACarSpring.businnes.concretes;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.etiya.rentACarSpring.businnes.abstracts.message.LanguageService;
+import com.etiya.rentACarSpring.businnes.abstracts.BrandService;
+import com.etiya.rentACarSpring.businnes.abstracts.CityService;
+import com.etiya.rentACarSpring.businnes.abstracts.ColorService;
 import com.etiya.rentACarSpring.businnes.abstracts.message.LanguageWordService;
+import com.etiya.rentACarSpring.businnes.abstracts.constants.Messages;
+import com.etiya.rentACarSpring.core.utilities.businnessRules.BusinnessRules;
 import com.etiya.rentACarSpring.core.utilities.results.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
@@ -30,16 +34,22 @@ public class CarManager implements CarService {
     private findexScoreService findexScoreService;
     private Environment environment;
     private LanguageWordService languageWordService;
+    private CityService cityService;
+    private BrandService brandService;
+    private ColorService colorService;
 
     @Autowired
     public CarManager(CarDao carDao, ModelMapperService modelMapperService, findexScoreService findexScoreService, Environment environment
-            ,LanguageWordService languageWordService) {
+            ,LanguageWordService languageWordService,CityService cityService, BrandService brandService, ColorService colorService) {
         super();
         this.carDao = carDao;
         this.modelMapperService = modelMapperService;
         this.findexScoreService = findexScoreService;
         this.environment = environment;
         this.languageWordService = languageWordService;
+        this.cityService=cityService;
+        this.colorService=colorService;
+        this.brandService=brandService;
 
     }
 
@@ -49,30 +59,49 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = result.stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response);
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarListed));
     }
 
     @Override
     public Result save(CreateCarRequest createCarRequest) {
+        Result result = BusinnessRules.run( cityService.checkIfCityExists(createCarRequest.getCityId()),
+                colorService.checkIfColorExists(createCarRequest.getColorId()),
+                brandService.checkIfBrandExists(createCarRequest.getBrandId())
+        );
+        if (result != null) {
+            return result;
+        }
         Car car = modelMapperService.forRequest().map(createCarRequest, Car.class);
         car.setFindexScore(findexScoreService.sendCarFindexScore());
         this.carDao.save(car);
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(1,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarAdded));
     }
 
     @Override
     public Result update(UpdateCarRequest updateCarRequest) {
+        Result result = BusinnessRules.run( checkCarExistsInGallery(updateCarRequest.getCarId()),
+                cityService.checkIfCityExists(updateCarRequest.getCityId()),
+                colorService.checkIfColorExists(updateCarRequest.getColorId()),
+                brandService.checkIfBrandExists(updateCarRequest.getBrandId())
+        );
+        if (result != null) {
+            return result;
+        }
         Car car = modelMapperService.forRequest().map(updateCarRequest, Car.class);
+        car.setFindexScore(this.carDao.getById(updateCarRequest.getCarId()).getFindexScore());
         this.carDao.save(car);
-        return new SuccesResult();
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarUpdated));
     }
 
     @Override
     public Result delete(DeleteCarRequest deleteCarRequest) {
-
+        Result result = BusinnessRules.run( checkCarExistsInGallery(deleteCarRequest.getCarId())
+        );
+        if (result != null) {
+            return result;
+        }
         this.carDao.deleteById(deleteCarRequest.getCarId());
-        return new SuccesResult();
-
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarDeleted));
     }
 
     @Override
@@ -81,7 +110,7 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = this.carDao.getByDailyPrice(dailyPrice).stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response);
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByDailyPriceListed));
     }
 
     @Override
@@ -89,7 +118,7 @@ public class CarManager implements CarService {
 
         List<CarDetail> response = this.carDao.getCarWithBrandAndColorDetails();
 
-        return new SuccesDataResult<List<CarDetail>>(response);
+        return new SuccesDataResult<List<CarDetail>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarDetailedListed));
     }
 
     @Override
@@ -98,7 +127,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response);
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByColorIdListed));
     }
 
     @Override
@@ -107,7 +136,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response);
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByBrandIdListed));
     }
 
     @Override
@@ -116,7 +145,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response);
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCarIdListed));
     }
 
     @Override
@@ -129,7 +158,7 @@ public class CarManager implements CarService {
         List<CarSearchListDto> result = this.carDao.getAllWithoutMaintenanceOfCar();
         List<CarSearchListDto> response = result.stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
-        return new SuccesDataResult<List<CarSearchListDto>>(response);
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarWithoutMaintenanceListed));
     }
 
     @Override
@@ -137,15 +166,15 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = this.carDao.getByCity_CityId(cityId).stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response);
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCityIdListed));
     }
 
     @Override
     public Result checkCarExistsInGallery(int id) {
-        if (this.carDao.existsById(id)) {
-            return new SuccesResult();
+        if (!this.carDao.existsById(id)) {
+            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarNotFound));
         }
-        return new ErrorResult();
+        return new SuccesResult();
     }
 
 }
